Allow setting of multiple request headers and cookies in a single step
The steps currently available require a separate step for each request
header or cookie that is set. This commit adds new steps that support
one or more cookies and one or more request headers being set in a single
step.

To avoid confusion I've also recommended that the steps for setting a
single header or cookie are deprecated and replaced completely with these
new steps.