fix: Cookies not persisted when using {base} variable in url
Superagent sets the shared cookies on the agent based on the domain
name of the request url. Currently the framework replaces a `{base}`
type var in the url too late - meaning that superagent doesn't see
the domain and set the cookies.

This changed uses `world.replaceVars` immediately to fix this issue