feat: Add "Then I should receive the text:" step
Asserts that a response body matches the expected text. Both strings
are trimmed. Also, this will not attempt to parse text. So JSON is a string
of text just the same at content-type: "text/html".

This commit also required a bit of refactoring to `world.js` and now the
`getResponse()` method is easier to work with and consistenly handles
error responses.