feat: Generalized support for better authentication
Added the following new steps:

1. `/^I am an? "([^"]*)"$/` E.g. "Given I am an "admin".

...which will explicity create a SuperAgent agent that can be reused
across scenarios. This is useful so that actions like requesting an oauth2
bearer token are repeated when they only need to be performed onnce.

This also is part of an effort to simplify code responsible for setting and
reusing agents. The previous approach tried to make it implicit, but I think
was confusing and in some scenarios not reliable due to the way we generated
the agent's dictionary key.

2. `I am using basic authentication using credentials from: {string}` and
`I am using basic authentication with the credentials:` plus short forms.

Adds support for Basic auth