fix: allow for repeated query parameters.
Currently the following case is not covered:

```
When qs:
      | select          | fullName |
      | select          | score    |
      | select          | created  |
      | select          | source   |
```

Only the last `select` parameter will be passed to the request.

This change will take the raw values from the query-string table and create arrays for duplicate keys.

I also changed to step for `print the request` to show query-string and replaced vars